+++ src/Vector/cuda/vector_dist_gpu_MP_tests.cu
@@ -523,25 +523,25 @@ BOOST_AUTO_TEST_CASE( vector_dist_multiphase_kernel_test )
 		phases.template get<0>(2).add();
 		phases.template get<0>(3).add();
 
-		phases.template get<0>(0).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.template get<0>(0).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.template get<0>(0).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.template get<0>(0).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.template get<0>(0).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(0).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(0).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(0).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 
-		phases.template get<0>(1).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.template get<0>(1).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.template get<0>(1).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.template get<0>(1).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.template get<0>(1).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(1).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(1).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(1).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 
-		phases.template get<0>(2).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.template get<0>(2).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.template get<0>(2).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.template get<0>(2).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.template get<0>(2).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(2).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(2).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(2).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 
-		phases.template get<0>(3).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.template get<0>(3).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.template get<0>(3).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.template get<0>(3).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.template get<0>(3).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(3).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(3).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.template get<0>(3).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 	}
 
 	phases.template get<0>(0).hostToDevicePos();
@@ -615,25 +615,25 @@ BOOST_AUTO_TEST_CASE( vector_dist_multiphase_kernel_test_simplified )
 		phases.get(2).add();
 		phases.get(3).add();
 
-		phases.get(0).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.get(0).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.get(0).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.get(0).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.get(0).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.get(0).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.get(0).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.get(0).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 
-		phases.get(1).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.get(1).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.get(1).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.get(1).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.get(1).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.get(1).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.get(1).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.get(1).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 
-		phases.get(2).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.get(2).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.get(2).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.get(2).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.get(2).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.get(2).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.get(2).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.get(2).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 
-		phases.get(3).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.get(3).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.get(3).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.get(3).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.get(3).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.get(3).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.get(3).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.get(3).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 	}
 
 	phases.get(0).hostToDevicePos();
@@ -706,25 +706,25 @@ BOOST_AUTO_TEST_CASE( vector_dist_multiphase_kernel_cl_test )
 		phases.get(2).add();
 		phases.get(3).add();
 
-		phases.get(0).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.get(0).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.get(0).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.get(0).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.get(0).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.get(0).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.get(0).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.get(0).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 
-		phases.get(1).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.get(1).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.get(1).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.get(1).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.get(1).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.get(1).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.get(1).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.get(1).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 
-		phases.get(2).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.get(2).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.get(2).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.get(2).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.get(2).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.get(2).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.get(2).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.get(2).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 
-		phases.get(3).getLastPosWrite()[0] = (float)rand() / RAND_MAX;
-		phases.get(3).getLastPosWrite()[1] = (float)rand() / RAND_MAX;
-		phases.get(3).getLastPosWrite()[2] = (float)rand() / RAND_MAX;
-		phases.get(3).getLastPropWrite<0>() = (float)rand() / RAND_MAX;
+		phases.get(3).getLastPosWrite()[0] = (float)rand() / (float)RAND_MAX;
+		phases.get(3).getLastPosWrite()[1] = (float)rand() / (float)RAND_MAX;
+		phases.get(3).getLastPosWrite()[2] = (float)rand() / (float)RAND_MAX;
+		phases.get(3).getLastPropWrite<0>() = (float)rand() / (float)RAND_MAX;
 	}
 
 	// redistribute all
